% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_freely_additions.R
\name{by_sm}
\alias{by_sm}
\title{apply function to each select_multiple response individually}
\usage{
by_sm(x, FUN, ...)
}
\arguments{
\item{x}{a vector of select_multiple responses separated by a space " "}

\item{FUN}{the function to be applied}

\item{...}{further parameters passed to FUN}
}
\value{
each value in x is split into a vector on " " (space); the function in FUN is applied to each of these vectors. We return a vector of these results
}
\description{
apply function to each select_multiple response individually
}
