% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_freely.R
\name{compose_freely}
\alias{compose_freely}
\title{compose freely with a custom condition}
\usage{
compose_freely(.data, to, where.expression, questionnaire = NULL, ...)
}
\arguments{
\item{.data}{an ongoing recoding (see new_recoding())}

\item{to}{the value to set to}

\item{where}{straight up R code; evaluated in the namespace of the input data (see examples)}

\item{condition_name}{: give your condition a name; defaults to "free condition"}

\item{questionnaire:}{if you supply a questionnaire, you will be able to use the following functions within condition:

- question_is_skipped("variable_name")}
}
\description{
compose freely with a custom condition
}
