% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Create and Compute Composite Variable Statistics}
\usage{
compute(loc, df)
}
\arguments{
\item{loc}{List of items grouped by scale, created using compose function}

\item{df}{The data frame to add the composite variable to}
}
\value{
A data frame with the composite variable added and basic statistics
}
\description{
Create and Compute Composite Variable Statistics
}
\examples{
#Use data from the compose function
#testData <- compute(test$var1, testData)
}
