% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{end_composition}
\alias{end_composition}
\title{end composition}
\usage{
end_composition(.data)
}
\arguments{
\item{.data}{the ongoing composition}
}
\value{
data.frame with the newly composed variable(s)
}
\description{
end composition
}
\details{
discards all composition meta information
}
