% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{end_recoding}
\alias{end_recoding}
\title{turn active recoding back into a simple data frame}
\usage{
end_recoding(.data)
}
\arguments{
\item{.data}{the recoding (see ?new_recoding)}
}
\value{
the data as a regular data.frame (tibble), with the new recoded variable added. All meta information on the recoding process is discarded.
}
\description{
turn active recoding back into a simple data frame
}
