% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lava}
\alias{lava}
\title{Skye lava}
\format{A data frame with twelve variables:
\describe{
\item{\code{id}}{unique sample id}
\item{\code{type}}{type of lava from either: Basalt, Hawaiite, or Benmoreite}
\item{\code{SiO2}}{proportion of silicon dioxide}
\item{\code{Al2O3}}{proportion of aluminium oxide}
\item{\code{Fe2O3}}{proportion of ferric oxide}
\item{\code{MgO}}{proportion of magnesium oxide}
\item{\code{CaO}}{proportion of calcium oxide}
\item{\code{Na2O}}{proportion of sodium oxide}
\item{\code{K2O}}{proportion of potassium oxide}
\item{\code{TiO2}}{proportion of titanium dioxide}
\item{\code{P2O3}}{proportion of phosphorus trioxide}
\item{\code{MnO}}{proportion of manganese(II) oxide}
}}
\usage{
lava
}
\description{
Chemical analyses of Skye lavas copied directly from table 2 of \insertCite{Thompson1972-wu}{composr}.
A quick summary will show the proportions across a sample do not sum to 100 but they have been copied correctly.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
