% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{new_composition}
\alias{new_composition}
\title{Start a new composition}
\usage{
new_composition(df, target)
}
\arguments{
\item{df}{the source data as a data.frame}

\item{target}{the name of the variable that will be composed and added to the data}
}
\value{
the input data frame with
 - an additional column named after the value of `target`
 - background setup to manage step by step composition of that variable from others.
}
\description{
Start a new composition
}
