% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{new_recoding}
\alias{new_recoding}
\title{Start a new recoding}
\usage{
new_recoding(df, source, target)
}
\arguments{
\item{df}{the source data as a data.frame}

\item{source}{the variable to recode from}

\item{target}{the name of the new variable created through the recoding}
}
\value{
the input data frame with
 - an additional column named after the value of `target`
 - background setup to manage step by step recoding of the source variable
}
\description{
Start a new recoding
}
\details{
When conditions are conflicting, the last condition that applies is used
 recoding is a special case of a composition, where the source variable is defined from the start and does not change.
}
