% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pebble}
\alias{pebble}
\title{Pebble}
\format{A data frame with seven variables:
\describe{
\item{\code{id}}{unique sample id}
\item{\code{type}}{sample from either (P)ebble, (T)ill, or (S)and sample}
\item{\code{red}}{percentile of red sandstone}
\item{\code{gray}}{percentile of gray sandstone}
\item{\code{cryst}}{percentile of crystalline rock}
\item{\code{misc}}{percentile of pebbles unable to be classified within other three groups}
\item{\code{n_pebbles}}{number of pebbles in each sample}
}}
\usage{
pebble
}
\description{
Data from pebble analyses copied directly from table 2 in \insertCite{Kaiser1962-dm}{composr}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
