% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_batch.R
\name{recode_batch}
\alias{recode_batch}
\title{apply many recodings at once with vector of 'where' conditions}
\usage{
recode_batch(df, tos, wheres, targets = NULL, questionnaire = NULL)
}
\arguments{
\item{df}{a data frame or an ongoing recoding}

\item{tos}{a vector of "to" values}

\item{wheres}{a vector of "where" conditions; R code as strings (evaluated in namespace of the data)}

\item{targets}{vector of target variables to create as characters. each change triggers a new_recoding(). if left empty, recodes to taret specified in new_recoding().}

\item{return}{the ongoing recoding from after the last 'where' recoding. return to regular data frame with all new recodings visible with end_recoding()}
}
\description{
apply many recodings at once with vector of 'where' conditions
}
