% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_directly.R
\name{recode_directly}
\alias{recode_directly}
\title{recode directly to a value}
\usage{
recode_directly(.data, to_expression, questionnaire = NULL, ...)
}
\arguments{
\item{.data}{an ongoing recoding (see new_recoding())}

\item{to_expression}{R code as a character string; evaluated in the namespace of the input data, and result will be the 'to' value; will overwrite everything that is not NA here}

\item{questionnaire}{if you supply a questionnaire, you will be able to use the following functions within condition:}
}
\description{
recode directly to a value
}
