% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sediment}
\alias{sediment}
\title{Sediment}
\format{A data frame with eight variables:
\describe{
\item{\code{id}}{unique station number}
\item{\code{depth}}{water depth in metres}
\item{\code{mean}}{graphic mean diameter of grain size}
\item{\code{sd}}{graphic standard deviation of grain size}
\item{\code{skew}}{graphic skewness of grain size}
\item{\code{sand}}{percentile of sand}
\item{\code{silt}}{percentile of silt}
\item{\code{clay}}{percentile of clay}
\item
}}
\usage{
sediment
}
\description{
Sedimentation analyses from an artic lake copied directly from table 1 in \insertCite{Coakley1968-eb}{composr}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
