% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settle.R
\name{settle}
\alias{settle}
\title{Lower Inflated Response Option Values}
\usage{
settle(df, points)
}
\arguments{
\item{df}{The data frame to examine}

\item{points}{Maximum number of response options}
}
\value{
A data frame with deflated response option values
}
\description{
Survey software will sometimes assign arbitrary values to response options. When this happens, a 1-5 scale may become a 14-18 scale. This function lowers all values with the specified range so that the minimum value is 1.
}
\examples{
#For seven-point scales: df1 <- settle(df, 7)
}
