% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_freely_additions.R
\name{sm_selected}
\alias{sm_selected}
\title{Check if select_multiple choices were selected}
\usage{
sm_selected(x, any = NULL, all = NULL, exactly = NULL, none = NULL)
}
\arguments{
\item{x}{a vector of select multiple responses, with choices separated by spaces}

\item{any}{TRUE if any of the values supplied here as a vector were selected}

\item{all}{TRUE if all of the values supplied here as a vector were selected}

\item{exactly}{TRUE if exactly all of the values supplied here as a vector were selected (an no others)}

\item{any}{TRUE if none of the values supplied here as a vector were selected}
}
\value{
a logical vector, same length as x
}
\description{
Check if select_multiple choices were selected
}
\details{
only supply one of any/all/exactly/any
}
