% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsettled.R
\name{unsettled}
\alias{unsettled}
\title{Identify Inflated Response Option Values}
\usage{
unsettled(df, points)
}
\arguments{
\item{df}{The data frame to examine}

\item{points}{Maximum number of response options}
}
\value{
A list of variables that may need manual deflation of response option values
}
\description{
For use in conjunction with settle(). This function identifies variables with outside of the specified range that may require manual recoding to address inflated response option values.
}
\examples{
#For seven-point scales: df1 <- unsettled(df, 7)
}
