% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Group Items into Scales}
\usage{
compose(df)
}
\arguments{
\item{df}{Data frame of items}
}
\value{
List of scales containing a list of items
}
\description{
Group Items into Scales
}
\examples{
#build test data
testData <- data.frame("Case" = matrix(1:10, ncol = 1, nrow = 10))
x = 1
while(x <= round(runif(1, 3, 6))){ #Number of Variables
  y = 1
  while(y <= round(runif(1, 3, 10))){ #Number of Items
    temp <- round(runif(10, 1, 7)) #Number of Responses
    testData <- cbind(testData, as.data.frame(temp))
    colnames(testData)[colnames(testData) == "temp"] <-
      paste0("Var", x, "_", y)
    y = y + 1
    }
  x = x + 1
}

#Group items into scales
test <- compose(testData)
}
