% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodings.R
\name{recode_to}
\alias{recode_to}
\title{add layer to current recoding}
\usage{
recode_to(.data, to, where.selected.any = NULL,
  where.selected.all = NULL, where.selected.exactly = NULL,
  where.selected.none = NULL, where.num.equal = NULL,
  where.num.smaller = NULL, where.num.smaller.equal = NULL,
  where.num.larger = NULL, where.num.larger.equal = NULL,
  otherwise.to = NA, skipped.to = NA, na.to = NA,
  questionnaire = NULL)
}
\arguments{
\item{.data}{the ongoing recoding obejct, see new_recoding()}

\item{to}{the value to set the new composition to if the condition is fulfilled}

\item{where.selected..}{: a vector of choices; setting values to 'to' where in the source variable any/all/exactly/none of the supplied choices had been selected}

\item{where.num...}{: a scalar number. setting values to 'to' where the 'source' is equal / smaller / smaller or equal / larger / larger or euqal than the number supplied in where.num...}

\item{otherwise}{an alternative value to be used if the condition is not fulfilled, the source is not NA and not skipped}
}
\value{
the updated composition
}
\description{
add layer to current recoding
}
